feat: break down Switch into smaller components
